feat(ui): connect dapp
