+++ src/locales/en/en.json
@@ -650,6 +650,66 @@
       }
     }
   },
+  "connectdapp": {
+    "title": "Cardano connect",
+    "sublabel": "CIP-45",
+    "tabheader": "Cardano Connect",
+    "connectbtn": "Connect with Cardano",
+    "disconnectbeforecreatealert": {
+      "message": "You are currently connected. To connect with a new wallet, you will be disconnected from your current selection. Would you like to disconnect and continue?",
+      "confirm": "Continue",
+      "cancel": "Cancel"
+    },
+    "connectionbrokenalert": {
+      "message": "Your connection has been disconnected as you have deleted the chosen identifier paired with this connection. Please choose a new identifier to re-establish this connection.",
+      "confirm": "Ok"
+    },
+    "inputpidmodal": {
+      "header": "Paste Meerkat ID",
+      "cancel": "Cancel",
+      "confirm": "Confirm",
+      "scanQR": "Scan QR code"
+    },
+    "connectionhistory": {
+      "title": "Connection History",
+      "action": {
+        "delete": "Delete"
+      },
+      "deletealert": {
+        "message": "Are you sure you want to delete this connection?",
+        "confirm": "Yes, I’m sure",
+        "cancel": "Cancel"
+      },
+      "missingidentifieralert": {
+        "message": "To connect with Cardano, you need to create an individual identifier. Would you like to create one now?",
+        "confirm": "Create identifier",
+        "cancel": "Cancel"
+      },
+      "confirmconnect": {
+        "done": "Done",
+        "connectbtn": "Connect",
+        "disconnectbtn": "Disconnect",
+        "connectingbtn": "Connecting...",
+        "pending": "Pending"
+      }
+    },
+    "request": {
+      "stageone": {
+        "title": "Connect Cardano",
+        "message": "Would you like to establish a CIP-45 Connection?",
+        "alert": {
+          "titleconfirm": "Are you sure you want to decline this connection?",
+          "confirm": "Decline",
+          "cancel": "Cancel"
+        }
+      },
+      "button": {
+        "back": "Back",
+        "accept": "Accept",
+        "decline": "Decline"
+      }
+    }
+  },
   "tabs": {
     "identifiers": {
       "layout": {
@@ -994,71 +1054,6 @@
           "connections": {
             "title": "Connections",
             "tabheader": "Connections"
-          },
-          "connectwallet": {
-            "title": "Cardano connect",
-            "sublabel": "CIP-45",
-            "tabheader": "Cardano Connect",
-            "connectbtn": "Connect with Cardano",
-            "disconnectbeforecreatealert": {
-              "message": "You are currently connected. To connect with a new wallet, you will be disconnected from your current selection. Would you like to disconnect and continue?",
-              "confirm": "Continue",
-              "cancel": "Cancel"
-            },
-            "connectionbrokenalert": {
-              "message": "Your connection has been disconnected as you have deleted the chosen identifier paired with this connection. Please choose a new identifier to re-establish this connection.",
-              "confirm": "Ok"
-            },
-            "inputpidmodal": {
-              "header": "Paste Meerkat ID",
-              "cancel": "Cancel",
-              "confirm": "Confirm",
-              "scanQR": "Scan QR code"
-            },
-            "connectionhistory": {
-              "title": "Connection History",
-              "action": {
-                "delete": "Delete"
-              },
-              "deletealert": {
-                "message": "Are you sure you want to delete this connection?",
-                "confirm": "Yes, I’m sure",
-                "cancel": "Cancel"
-              },
-              "missingidentifieralert": {
-                "message": "To connect with Cardano, you need to create an individual identifier. Would you like to create one now?",
-                "confirm": "Create identifier",
-                "cancel": "Cancel"
-              },
-              "confirmconnect": {
-                "done": "Done",
-                "connectbtn": "Connect",
-                "disconnectbtn": "Disconnect",
-                "connectingbtn": "Connecting...",
-                "pending": "Pending"
-              }
-            },
-            "request": {
-              "stageone": {
-                "title": "Connect Cardano",
-                "message": "Would you like to establish a CIP-45 Connection?",
-                "alert": {
-                  "titleconfirm": "Are you sure you want to decline this connection?",
-                  "confirm": "Decline",
-                  "cancel": "Cancel"
-                }
-              },
-              "stagetwo": {
-                "title": "Choose identifier",
-                "message": "Which individual identifier would you like to make the connection with?",
-                "confirm": "Confirm"
-              },
-              "button": {
-                "back": "Back",
-                "accept": "Accept",
-                "decline": "Decline"
-              }
-            }
           }
         }
       }
