+++ src/ui/components/ConnectdApp/ConnectdApp.scss
@@ -1,3 +1,14 @@
+.wallet-connect {
+  & .page-header {
+    ion-toolbar {
+      ion-button.close-button,
+      ion-button.connect-wallet-button {
+        background: var(--ion-color-neutral-100);
+      }
+    }
+  }
+}
+
 .connect-wallet-container {
   .connect-wallet-title {
     font-size: 1.25rem;
