+++ src/ui/components/ConnectdApp/ConnectdApp.types.ts
@@ -1,0 +1,22 @@
+import { ConnectionData } from "../../../store/reducers/walletConnectionsCache";
+
+interface ConnectdAppProps {
+  isOpen: boolean;
+  setIsOpen: (value: boolean) => void;
+}
+
+type ActionInfo = {
+  type: ActionType;
+  data?: ConnectionData;
+};
+
+enum ActionType {
+  Add = "add",
+  Delete = "delete",
+  Connect = "connect",
+  None = "none",
+}
+
+export { ActionType };
+
+export type { ActionInfo, ConnectdAppProps, ConnectionData };
