+++ src/ui/components/ConnectdApp/components/ConfirmConnectModal/ConfirmConnectModal.scss
