+++ src/ui/components/ConnectdApp/components/ConfirmConnectModal/ConfirmConnectModal.types.ts
