+++ src/ui/components/ConnectdApp/components/ConfirmConnectModal/index.ts
