+++ src/ui/components/ConnectdApp/components/WalletConnect/index.ts
