+++ src/ui/components/ConnectdApp/index.ts
@@ -1,0 +1,2 @@
+export * from "./ConnectdApp";
+export * from "./ConnectdApp.types";
