--- src/ui/pages/Menu/Menu.scss
@@ -1,69 +1,0 @@
-.menu-tab {
-  ion-header {
-    margin-bottom: 1rem;
-  }
-
-  ion-content {
-    ion-row {
-      margin: 0 -0.3rem;
-
-      ion-col {
-        .menu-input {
-          margin: 0 0.3rem;
-          display: flex;
-          background: var(--ion-color-neutral-100);
-          flex-direction: column;
-          justify-content: space-evenly;
-          height: 5.875rem;
-          padding: 1.25rem;
-          box-shadow: none;
-          --inner-border-width: 0;
-          color: var(--ion-color-neutral-700);
-
-          ion-icon {
-            font-size: 1.25rem;
-          }
-        }
-
-        .menu-item-icon {
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-
-          .sub-label {
-            padding: 0.125rem 0.375rem;
-            border-radius: 0.25rem;
-            background-color: var(--ion-color-neutral-300);
-            font-size: 0.5rem;
-            font-weight: 400;
-            line-height: 0.75rem;
-          }
-        }
-      }
-    }
-  }
-
-  @media screen and (min-width: 250px) and (max-width: 370px) {
-    ion-header {
-      margin-bottom: 0.8rem;
-    }
-
-    ion-content {
-      ion-row {
-        margin: 0 -0.3rem;
-
-        ion-col {
-          .menu-input {
-            margin: 0 0.3rem;
-            height: 4.7rem;
-            padding: 1rem;
-
-            ion-icon {
-              font-size: 1rem;
-            }
-          }
-        }
-      }
-    }
-  }
-}
