--- src/ui/pages/Menu/components/ConnectWallet/ConnectWallet.types.ts
@@ -1,21 +1,0 @@
-import { ConnectionData } from "../../../../../store/reducers/walletConnectionsCache";
-
-interface ConnectWalletOptionRef {
-  openConnectWallet: () => void;
-}
-
-type ActionInfo = {
-  type: ActionType;
-  data?: ConnectionData;
-};
-
-enum ActionType {
-  Add = "add",
-  Delete = "delete",
-  Connect = "connect",
-  None = "none",
-}
-
-export { ActionType };
-
-export type { ConnectWalletOptionRef, ActionInfo, ConnectionData };
