--- src/ui/pages/Menu/components/ConnectWallet/index.ts
@@ -1,2 +1,0 @@
-export * from "./ConnectWallet";
-export * from "./ConnectWallet.types";
