--- src/ui/pages/Menu/components/SubMenu/SubMenu.scss
@@ -1,87 +1,0 @@
-.sub-menu {
-  display: flex !important;
-
-  &.show {
-    transform: translate3d(0, 0, 0);
-    transition-duration: 0.5s;
-  }
-
-  &.hide {
-    transform: translate3d(100%, 0, 0);
-    transition-duration: 0.5s;
-  }
-
-  ion-header.page-header ion-toolbar {
-    margin-top: 0.625rem;
-    padding-right: 0;
-    padding-left: 0;
-  }
-
-  ion-toolbar ion-buttons ion-button.button-has-icon-only {
-    --background-activated: var(--ion-color-neutral-100);
-    --background-focused: var(--ion-color-neutral-100);
-    --background-hover: var(--ion-color-neutral-100);
-    background: var(--ion-color-neutral-100);
-  }
-
-  &.connections .cards-placeholder-container {
-    height: 86vh;
-  }
-
-  &.recovery-seed-phrase {
-    &.md {
-      padding-bottom: 0 !important;
-    }
-  }
-
-  &.recovery-seed-phrase .recovery-phrase-content {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    min-height: calc(100vh - (8.5rem + var(--ion-safe-area-top)));
-    margin-bottom: 0 !important;
-  }
-
-  &.connect-wallet {
-    background: var(--ion-color-neutral-200);
-
-    ion-header ion-toolbar {
-      --background: transparent;
-
-      ion-title.md {
-        margin-left: 0;
-      }
-    }
-
-    ion-toolbar ion-buttons ion-button.close-button {
-      --background-activated: var(--ion-color-neutral-100);
-      --background-focused: var(--ion-color-neutral-100);
-      --background-hover: var(--ion-color-neutral-100);
-      background: var(--ion-color-neutral-100);
-    }
-
-    ion-toolbar ion-buttons ion-button.connect-wallet-button {
-      --padding-start: 0.5rem;
-      --padding-end: 0.5rem;
-      --background-activated: var(--ion-color-neutral-100);
-      --background-focused: var(--ion-color-neutral-100);
-      --background-hover: var(--ion-color-neutral-100);
-      background: var(--ion-color-neutral-100);
-    }
-  }
-
-  @media screen and (min-width: 250px) and (max-width: 370px) {
-    ion-header ion-toolbar {
-      margin-top: 0.531rem;
-    }
-
-    &.connections .cards-placeholder-container {
-      height: 78vh;
-    }
-
-    ion-toolbar ion-buttons ion-button.connect-wallet-button {
-      --padding-start: 0.425rem;
-      --padding-end: 0.425rem;
-    }
-  }
-}
