--- src/ui/pages/Menu/components/SubMenu/index.ts
@@ -1,1 +1,0 @@
-export * from "./SubMenu";
