--- src/ui/pages/WalletConnect/WalletConnect.scss
@@ -1,108 +1,0 @@
-.wallet-connect-container {
-  width: 100%;
-  height: 100%;
-  background: var(--ion-color-neutral-200);
-}
-
-.connect-wallet-stage-one {
-  & .page-header {
-    ion-toolbar {
-      --background: transparent;
-    }
-  }
-
-  .responsive-page-content {
-    --buttons-height: 7.5rem;
-    justify-content: space-between;
-    height: 100%;
-
-    .request-animation-center {
-      flex: 1;
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-
-      .request-icons-row {
-        display: flex;
-        justify-content: center;
-        margin-bottom: 1.5rem;
-
-        .request-user-logo {
-          border-radius: 5rem;
-          background-color: var(--ion-color-neutral-700);
-          width: 4.875rem;
-          height: 4.875rem;
-
-          & > ion-icon {
-            width: 100%;
-            height: 100%;
-          }
-        }
-
-        @media screen and (min-width: 250px) and (max-width: 370px) {
-          .request-user-logo {
-            width: 3.5rem;
-            height: 3.5rem;
-          }
-        }
-      }
-
-      .wallet-connect-message {
-        text-align: center;
-        margin: 0;
-      }
-    }
-  }
-}
-
-.wallet-connect-stage-two {
-  & .page-header {
-    ion-toolbar {
-      --background: transparent;
-    }
-  }
-
-  .responsive-page-content {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    .title {
-      font-size: 1rem;
-      line-height: 1.115rem;
-      margin-top: 0.75rem;
-      margin-bottom: 1.75rem;
-      font-weight: 500;
-      text-align: center;
-    }
-
-    .identifier-list {
-      .checkbox {
-        flex: none;
-        width: fit-content;
-        --checkbox-background-checked: var(--ion-color-primary-700);
-        --checkmark-color: var(--ion-color-primary-100);
-        --border-width: 0.115rem;
-
-        &.checkbox-checked {
-          --border-width: 0;
-        }
-      }
-
-      .card-item {
-        --background-activated: transparent;
-      }
-    }
-
-    @media screen and (min-width: 250px) and (max-width: 370px) {
-      .title {
-        font-size: 0.85rem;
-        margin-bottom: 1.5rem;
-      }
-
-      .checkbox {
-        --size: 1.25rem;
-      }
-    }
-  }
-}
